requirements.txt
Flask==3.0.3
feedparser==6.0.11
scikit-learn==1.5.1
numpy==1.26.4
google-generativeai==0.7.2
gunicorn==22.0.0
